Completed API and configuration

+++ src/tasks/tasks.service.ts
@@ -5,6 +5,7 @@ import { GetTasksDto } from './dto/get-tasks.dto';
 import { Task } from './task.entity';
 import { InjectRepository } from '@nestjs/typeorm';
 import { TaskRepository } from './task.repository';
+import { User } from '../auth/user.entity';
 
 @Injectable()
 export class TasksService {
@@ -12,12 +13,12 @@ export class TasksService {
     @InjectRepository(TaskRepository) private taskRepo: TaskRepository,
   ) {}
 
-  async getTasks(getTasksDto: GetTasksDto): Promise<Task[]> {
-    return this.taskRepo.getTasks(getTasksDto);
+  async getTasks(getTasksDto: GetTasksDto, user: User): Promise<Task[]> {
+    return this.taskRepo.getTasks(getTasksDto, user);
   }
 
-  async getTaskById(id: number): Promise<Task> {
-    const task = await this.taskRepo.findOne(id);
+  async getTaskById(id: number, user: User): Promise<Task> {
+    const task = await this.taskRepo.findOne({ where: { id, userId: user.id } });
 
     if (!task) {
       throw new NotFoundException(`Task with ID ${id} not found`);
@@ -26,12 +27,12 @@ export class TasksService {
     return task;
   }
 
-  createTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepo.createTask(createTaskDto);
+  createTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepo.createTask(createTaskDto, user);
   }
 
-  async deleteTask(id: number): Promise<number> {
-    const deletedTask = await this.taskRepo.delete(id);
+  async deleteTask(id: number, user: User): Promise<number> {
+    const deletedTask = await this.taskRepo.delete({ id, userId: user.id });
 
     if (!deletedTask.affected) {
       throw new NotFoundException(`Task with ID ${id} not found`);
@@ -40,8 +41,8 @@ export class TasksService {
     return id;
   }
 
-  async updateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskById(id);
+  async updateTaskStatus(id: number, status: TaskStatus, user): Promise<Task> {
+    const task = await this.getTaskById(id, user);
     task.status = status;
     await task.save();
     return task;
